非JSMS仕様のメモリレジスタ初期値を0に、ATS起動ロジック修正など

+++ dev.py
@@ -55,7 +55,8 @@ for folder in TARGET_FOLDERS:
                     dirs_exist_ok=True
                 )
 
-shutil.rmtree(dist_path)
+if dist_path.is_dir():
+    shutil.rmtree(dist_path)
 
 subprocess.run(root_path.joinpath("build.bat"), check=True)
